Change how constants are read from dat, reformat
scripts are now in the root directory, constants are directly `eval`'d
as such, satellite just.... works? except for some numerical errors and
weird printing.

+++ test/runtests.jl
@@ -4,7 +4,7 @@ using Test
 const pi = gps.pi
 const π = gps.pi
 # lamp post b12
-const b12 = (dms2rad(40,45,55.0),-dms2rad(111,50,58.0), 1372.)
+const b12 = (dms2rad(40,45,55.0),-dms2rad(111,50,58.0), BigFloat(1372.))
 @testset "gps.jl" begin
     include("prelude.jl")
     include("satellites.jl")
